+++ samples/ecommerce/build.gradle
@@ -6,7 +6,7 @@ plugins {
 
 android {
     namespace 'io.github.sceneview.sample.ecommerce'
-    compileSdk 33
+    compileSdk 34
 
     defaultConfig {
         applicationId "io.github.sceneview.sample.ecommerce"
