+++ samples/ecommerce/src/main/AndroidManifest.xml
@@ -6,12 +6,12 @@
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:supportsRtl="true"
-        android:theme="@style/Theme.SceneView">
+        android:theme="@style/Theme.SceneViewSample">
         <activity
             android:name=".MainActivity"
             android:exported="true"
             android:label="@string/app_name"
-            android:theme="@style/Theme.SceneView">
+            android:theme="@style/Theme.SceneViewSample">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
