+++ samples/ecommerce/src/main/java/io/github/sceneview/sample/ecommerce/productdescription/presentation/ProductDescriptionScreen.kt
@@ -15,12 +15,14 @@ import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.compose.ui.zIndex
 import androidx.navigation.NavHostController
+import io.github.sceneview.sample.ecommerce.R
 import java.math.BigDecimal
 
 @Composable
@@ -41,7 +43,7 @@ fun ProductDescriptionScreen(
         // UI actions must be run once, thus run them in a Launched Effect
         is ProductDescriptionUIAction.NavigateToAddToCartScreen -> {
             LaunchedEffect(Unit) {
-                Toast.makeText(context, "Added to cart!", Toast.LENGTH_SHORT).show()
+                Toast.makeText(context, context.getString(R.string.added_to_cart), Toast.LENGTH_SHORT).show()
                 productViewModel.onConsumeUIAction()
             }
         }
@@ -67,7 +69,7 @@ fun ProductDescriptionScreen(
                     }, modifier = Modifier.align(Alignment.CenterHorizontally),
                     shape = RoundedCornerShape(50)
                 ) {
-                    Text("Try-on Virtually")
+                    Text(stringResource(R.string.view_in_your_space))
                 }
 
                 Column(
@@ -104,7 +106,7 @@ fun ProductDescriptionScreen(
                         }, modifier = Modifier.weight(1f),
                         shape = RoundedCornerShape(50)
                     ) {
-                        Text("Add To Cart", textAlign = TextAlign.Center)
+                        Text(stringResource(R.string.add_to_cart), textAlign = TextAlign.Center)
                     }
                 }
             }
