--- samples/ecommerce/src/main/java/io/github/sceneview/sample/ecommerce/ui/theme/Shape.kt
@@ -1,11 +1,0 @@
-package io.github.sceneview.sample.ecommerce.ui.theme
-
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.Shapes
-import androidx.compose.ui.unit.dp
-
-val Shapes = Shapes(
-    small = RoundedCornerShape(4.dp),
-    medium = RoundedCornerShape(4.dp),
-    large = RoundedCornerShape(0.dp)
-)
